move wowless.db2 to tools.db2
