+++ tools/db2.lua
