+++ tools/db2dump.lua
@@ -5,7 +5,7 @@ local args = (function()
   parser:argument('db2', 'db2')
   return parser:parse()
 end)()
-local rows = require(args.dbc and 'tools.dbc' or 'wowless.db2').rows
+local rows = require(args.dbc and 'tools.dbc' or 'tools.db2').rows
 local dbdefs = dofile('build/products/' .. args.product .. '/dbdefs.lua')
 local db2file = 'extracts/' .. args.product .. '/db2/' .. args.db2 .. '.db2'
 local content = assert(require('pl.file').read(db2file))
